+++ RELEASE_NOTES.md
@@ -3,6 +3,7 @@
 ## New in git-machete 3.36.2
 
 - improved: formatting of the output of `git machete ... --help` and `git machete help ...`
+- fixed: `git machete -v` (no command) crashing
 
 ## New in git-machete 3.36.1